git-svn-id: https://svn.kuali.org/repos/kpme/branches/version-rice-2-0-br@2951 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ src/main/java/org/kuali/hr/lm/leavecalendar/web/LeaveCalendarAction.java
@@ -6,6 +6,7 @@ import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 import org.joda.time.DateTime;
+import org.kuali.hr.lm.LMConstants;
 import org.kuali.hr.lm.leaveblock.LeaveBlock;
 import org.kuali.hr.lm.leavecalendar.LeaveCalendarDocument;
 import org.kuali.hr.lm.leaveplan.LeavePlan;
@@ -92,6 +93,7 @@ public class LeaveCalendarAction extends TkAction {
 		
 		if (lcd != null) {
 			setupDocumentOnFormContext(lcf, lcd);
+            lcdh = lcd.getLeaveCalendarDocumentHeader();
 		} else {
 			LOG.error("Null leave calendar document in LeaveCalendarAction.");
 		}
@@ -109,9 +111,14 @@ public class LeaveCalendarAction extends TkAction {
 		lcf.setLeaveCalendar(calendar);
 		
 		this.populateCalendarAndPayPeriodLists(request, lcf);
-		
 		// KPME-1447
-		List<LeaveBlock> leaveBlocks = TkServiceLocator.getLeaveBlockService().getLeaveBlocksForDocumentId(lcd.getDocumentId());
+		//List<LeaveBlock> leaveBlocks = TkServiceLocator.getLeaveBlockService().getLeaveBlocksForDocumentId(lcd.getDocumentId());
+        List<LeaveBlock> leaveBlocks;
+        if (lcdh != null) {
+            leaveBlocks = TkServiceLocator.getLeaveBlockService().getLeaveBlocks(lcdh.getPrincipalId(), lcdh.getBeginDate(), lcdh.getEndDate());
+        } else {
+            leaveBlocks = Collections.emptyList();
+        }
         LeaveCalendar leaveCalender = new LeaveCalendar(viewPrincipal, calendarEntry);
         LeaveBlockAggregate aggregate = new LeaveBlockAggregate(leaveBlocks, calendarEntry, leaveCalender);
         lcf.setLeaveBlockString(LeaveActionFormUtils.getLeaveBlocksJson(aggregate.getFlattenedLeaveBlockList()));
@@ -172,7 +179,7 @@ public class LeaveCalendarAction extends TkAction {
 		Assignment assignment = TkServiceLocator.getAssignmentService().getAssignment(lcd, lcf.getSelectedAssignment());
 
 		TkServiceLocator.getLeaveBlockService().addLeaveBlocks(beginDate,
-				endDate, lcf.getCalendarEntry(), selectedEarnCode, hours, desc, assignment, spanningWeeks); // KPME-1446
+				endDate, lcf.getCalendarEntry(), selectedEarnCode, hours, desc, assignment, spanningWeeks, LMConstants.LEAVE_BLOCK_TYPE.LEAVE_CALENDAR); // KPME-1446
 		// after adding the leave block, set the fields of this form to null for future new leave blocks
 		lcf.setLeaveAmount(null);
 		lcf.setDescription(null);
@@ -186,9 +193,12 @@ public class LeaveCalendarAction extends TkAction {
 		LeaveCalendarForm lcf = (LeaveCalendarForm) form;
 		String leaveBlockId = lcf.getLeaveBlockId();
 
-		// TODO: need security check
-		TkServiceLocator.getLeaveBlockService().deleteLeaveBlock(
+        LeaveBlock blockToDelete = TkServiceLocator.getLeaveBlockService().getLeaveBlock(new Long(leaveBlockId));
+        if (blockToDelete != null
+                && TkServiceLocator.getPermissionsService().canDeleteLeaveBlock(blockToDelete)) {
+		    TkServiceLocator.getLeaveBlockService().deleteLeaveBlock(
 				Long.parseLong(leaveBlockId));
+        }
 
 		return mapping.findForward("basic");
 	}
@@ -201,23 +211,24 @@ public class LeaveCalendarAction extends TkAction {
 		
 		LeaveBlock updatedLeaveBlock = null;
 		updatedLeaveBlock = TkServiceLocator.getLeaveBlockService().getLeaveBlock(new Long(leaveBlockId));
-		if (StringUtils.isNotBlank(lcf.getDescription())) {
-			updatedLeaveBlock.setDescription(lcf.getDescription().trim());	
-    	}
-    	if (!updatedLeaveBlock.getLeaveAmount().equals(lcf.getLeaveAmount())) {
-    		updatedLeaveBlock.setLeaveAmount(lcf.getLeaveAmount());
-    	}
-    	EarnCode earnCode =  TkServiceLocator.getEarnCodeService().getEarnCodeById(selectedEarnCode); // selectedEarnCode = hrEarnCodeId
-		if (!updatedLeaveBlock.getEarnCode().equals(earnCode.getEarnCode())) {
-    		updatedLeaveBlock.setEarnCode(earnCode.getEarnCode());
-    		// update hr_earn_code_id as well
-    		updatedLeaveBlock.setEarnCodeId(selectedEarnCode);
-    	}
-		TkServiceLocator.getLeaveBlockService().updateLeaveBlock(updatedLeaveBlock);
-		lcf.setLeaveAmount(null);
-		lcf.setDescription(null);
-		lcf.setSelectedEarnCode(null);
-		
+        if (updatedLeaveBlock.isEditable()) {
+            if (StringUtils.isNotBlank(lcf.getDescription())) {
+                updatedLeaveBlock.setDescription(lcf.getDescription().trim());
+            }
+            if (!updatedLeaveBlock.getLeaveAmount().equals(lcf.getLeaveAmount())) {
+                updatedLeaveBlock.setLeaveAmount(lcf.getLeaveAmount());
+            }
+            EarnCode earnCode =  TkServiceLocator.getEarnCodeService().getEarnCodeById(selectedEarnCode); // selectedEarnCode = hrEarnCodeId
+            if (!updatedLeaveBlock.getEarnCode().equals(earnCode.getEarnCode())) {
+                updatedLeaveBlock.setEarnCode(earnCode.getEarnCode());
+                // update hr_earn_code_id as well
+                updatedLeaveBlock.setEarnCodeId(selectedEarnCode);
+            }
+            TkServiceLocator.getLeaveBlockService().updateLeaveBlock(updatedLeaveBlock);
+            lcf.setLeaveAmount(null);
+            lcf.setDescription(null);
+            lcf.setSelectedEarnCode(null);
+        }
         return mapping.findForward("basic");
     }
